Ensure the lexer channel is drained properly.
The lexer channel was not being read to completion in many cases and
the goroutine feeding the channel was left hanging indefinitely.
Added a utility function to read from the channel until exhaustion
and adding a comment that all callers of lex() and friends should
use 'defer' to invoke the drain function at some point.