Remove Tcl interpreter, move Scheme interpreter
The Scheme interpreter has been moved to its own GitHub project,
named bakeneko (https://github.com/nlfiedler/bakeneko).

The Tcl interpreter is being abandoned because it is hopelessly
difficult to create a reasonable Tcl interpreter. Best to focus
energy on the Scheme interpreter, one because it is a much simpler
language, and two because it is far more powerful.